Move pretrain_ under examples
