+++ examples/classes/pretrain_vision_classify.py
