+++ examples/inference/t5/simple_t5_batch_inference.py
@@ -6,7 +6,7 @@ from argparse import Namespace
 
 import torch
 
-import pretrain_t5
+import examples.classes.pretrain_t5 as pretrain_t5
 from megatron.core.inference.engines import AbstractEngine, StaticInferenceEngine
 from megatron.core.inference.inference_request import InferenceRequest
 from megatron.core.inference.model_inference_wrappers.inference_wrapper_config import (
@@ -21,7 +21,7 @@ from megatron.core.inference.text_generation_controllers.encoder_decoder_text_ge
 )
 from megatron.core.tokenizers.utils.build_tokenizer import build_tokenizer
 from megatron.core.transformer.module import MegatronModule
-from pretrain_t5 import model_provider
+from examples.classes.pretrain_t5 import model_provider
 
 sys.path.append(
     os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir, os.path.pardir))
