+++ examples/mixtral/train_mixtral_8x7b_distributed.sh
@@ -107,7 +107,7 @@ if [ -n "${WANDB_API_KEY}" ]; then
 fi
 
 
-torchrun ${DISTRIBUTED_ARGS[@]} pretrain_gpt.py \
+torchrun ${DISTRIBUTED_ARGS[@]} examples/classes/pretrain_gpt.py \
     ${MODEL_ARGS[@]} \
     ${MOE_ARGS[@]} \
     ${DATA_ARGS[@]} \
