+++ examples/t5/train_t5_220m_distributed.sh
@@ -69,7 +69,7 @@ OUTPUT_ARGS="
     --eval-iters 10
 "
 
-torchrun $DISTRIBUTED_ARGS pretrain_t5.py \
+torchrun $DISTRIBUTED_ARGS examples/classes/pretrain_t5.py \
     $T5_ARGS \
     $DATA_ARGS \
     $OUTPUT_ARGS \
