+++ tests/test_utils/recipes/h100/mamba.yaml
@@ -44,7 +44,7 @@ spec:
         "TENSORBOARD_PATH={assets_dir}/tensorboard"
         "CHECKPOINT_SAVE_PATH={artifacts_dir}/checkpoints"
         "CHECKPOINT_LOAD_PATH=/mnt/artifacts/model/{name}"
-        "TRAINING_SCRIPT_PATH=pretrain_mamba.py"
+        "TRAINING_SCRIPT_PATH=examples/classes/pretrain_mamba.py"
         "TRAINING_PARAMS_PATH=./tests/functional_tests/test_cases/{model}/{test_case}/model_config.yaml"
         "GOLDEN_VALUES_PATH=./tests/functional_tests/test_cases/{model}/{test_case}/golden_values_{environment}_{platforms}.json"
         "N_REPEAT={n_repeat}"
