+++ tools/bert_embedding/embed.py
@@ -19,7 +19,7 @@ from megatron.core.enums import ModelType
 from megatron.core.pipeline_parallel import get_forward_backward_func
 from megatron.legacy.model import BertModel
 from megatron.training.training import setup_model_and_optimizer
-from pretrain_bert import model_provider, get_batch, loss_func, forward_step
+from examples.classes.pretrain_bert import model_provider, get_batch, loss_func, forward_step
 
 from .dataset import BertEmbeddingDataset
 from .external_libs import h5py
