+++ tools/checkpoint/saver_core.py
@@ -44,7 +44,7 @@ class MegatronCheckpointSaverLLM(MegatronCheckpointSaverBase):
             self.model_provider = partial(model_provider, gpt_builder)
             self.margs.model_type = ModelType.encoder_or_decoder
         elif self.md.model_type == 'BERT':
-            from pretrain_bert import model_provider
+            from examples.classes.pretrain_bert import model_provider
             self.model_provider = model_provider
             self.margs.model_type = ModelType.encoder_or_decoder
         else:
